DeltaTime and Updates first steps
Actual Version: v1.0.0.2
Minor Changes: Improved the StaticElementsLayer drawing function: Now
there is no need for loading the sprite image at each step of the loop.
Major Changes: Splitted the drawing function into two: One for updating
the changes of the layer elements and another for the drawing.
Started the integration of the deltaTime concept for the gameLoop speed
consistency.

+++ Gems2D/Camera.h
@@ -60,8 +60,8 @@ class Camera {
 	int m_window_height;
 	int m_map_width;
 	int m_map_height;
-	int m_obs_x;
-	int m_obs_y;
+	float m_obs_x;
+	float m_obs_y;
 };
